core/requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Jinja2==3.1.2
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Job Queue
Celery==5.3.4
redis==5.0.1

# AI Integration
openai==1.3.7
anthropic==0.8.1

# Code Generation
black==23.11.0
isort==5.12.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
python-decouple==3.8